Use long for health, adjust scaling and animation
Change enemy health to 64-bit (BaseHealth and computed health) and update HealthHandler.TakeDamage to accept long to support larger HP values (EnemyBasic.cs, HealthHandler.cs). Tweak enemy health growth exponent from 1.2→1.15 to slow scaling. Enable AnimatedSprite2D autoplay and adjust frame_progress in EnemyShadow.tscn. Also slow shop upgrade cost growth by changing exponent from 0.3→0.22 in Upgrades.cs to rebalance progression.

+++ Enemy/EnemyBasic/HealthHandler.cs
@@ -31,7 +31,7 @@ public partial class HealthHandler : Node
         _progressBar.Value = CurrentHealth;
     }
     
-    public void TakeDamage(int damage)
+    public void TakeDamage(long damage)
     {
         CurrentHealth -= damage;
